second commit

+++ Account Microservice/Models/Statement.cs
@@ -14,8 +14,8 @@ namespace Account_Microservice.Models
         public DateTime date { get; set; }
         public string refno {get;set;}
         public DateTime ValueDate { get; set; }
-        public float Withdrawal { get; set; }
-        public float Deposit { get; set; }
-        public float ClosingBalance { get; set;}
+        public int Withdrawal { get; set; }
+        public int Deposit { get; set; }
+        public int ClosingBalance { get; set;}
     }
 }
