+++ Account Microservice/Models/TransactionStatus.cs
@@ -9,8 +9,8 @@ namespace Account_Microservice.Models
     {
         public string Message { get; set; }
 
-        public float source_balance { get; set; }
+        public int source_balance { get; set; }
 
-        public float destination_balance { get; set; }
+        public int destination_balance { get; set; }
     }
 }
